MaryiaF/refactor: convert Journal to TS
