+++ packages/bot-web-ui/src/components/journal/date-item/index.js
@@ -1,0 +1,3 @@
+import { DateItem } from './date-item';
+
+export default DateItem;
