+++ packages/bot-web-ui/src/components/journal/filter-dialog/filter-dialog.tsx
@@ -1,0 +1,29 @@
+import { useOnClickOutside } from '@deriv/components';
+import React from 'react';
+import { TFilterDialogProps } from '../journal.types';
+import Filters from '../filters';
+
+export const FilterDialog = ({
+    toggle_ref,
+    checked_filters,
+    filters,
+    filterMessage,
+    is_filter_dialog_visible,
+    toggleFilterDialog,
+}: TFilterDialogProps) => {
+    const wrapper_ref = React.useRef<HTMLDivElement>(null);
+    const validateClickOutside = (event: React.ChangeEvent<HTMLInputElement>) =>
+        is_filter_dialog_visible && !toggle_ref.current?.contains(event.target);
+
+    useOnClickOutside(wrapper_ref, toggleFilterDialog, validateClickOutside);
+
+    return (
+        <Filters
+            wrapper_ref={wrapper_ref}
+            checked_filters={checked_filters}
+            filters={filters}
+            filterMessage={filterMessage}
+            className='filter-dialog'
+        />
+    );
+};
