+++ packages/bot-web-ui/src/components/journal/filter-dialog/index.js
@@ -1,0 +1,3 @@
+import { FilterDialog } from './filter-dialog';
+
+export default FilterDialog;
