+++ packages/bot-web-ui/src/components/journal/filters/index.js
@@ -1,0 +1,3 @@
+import { Filters } from './filters.tsx';
+
+export default Filters;
