+++ packages/bot-web-ui/src/components/journal/format-message/index.js
@@ -1,0 +1,3 @@
+import { FormatMessage } from './format-message';
+
+export default FormatMessage;
