+++ packages/bot-web-ui/src/components/journal/index.js
@@ -1,4 +1,4 @@
-import Journal from './journal.jsx';
+import Journal from './journal';
 import './journal.scss';
 
 export default Journal;
