+++ packages/bot-web-ui/src/components/journal/journal-item/index.js
@@ -1,0 +1,3 @@
+import { JournalItem } from './journal-item';
+
+export default JournalItem;
