+++ packages/bot-web-ui/src/components/journal/journal-loader/index.js
@@ -1,0 +1,3 @@
+import { JournalLoader } from './journal-loader';
+
+export default JournalLoader;
