+++ packages/bot-web-ui/src/components/journal/journal-tools/index.js
@@ -1,0 +1,3 @@
+import { JournalTools } from './journal-tools';
+
+export default JournalTools;
